Fix negate unsigned warnings.
Unclear why these did not surface prior to rtc_feature merge.

Behaviour is well defined, but understandbly looks suspect, hence use of cast.

+++ include/flamegpu/runtime/messaging/Array.h
@@ -200,7 +200,7 @@ class MsgArray {
              */
             inline __device__ iterator begin(void) const {
                 // Bin before initial bin, as the constructor calls increment operator
-                return iterator(*this, -radius - 1);
+                return iterator(*this, -static_cast<int>(radius) - 1);
             }
             /**
              * Returns an iterator to the position beyond the end of the message list subset
